apply feedback
